Move csv route to /api/csv and add support for month/year filtering
